DP-8453: Prevent listing header from being loaded on every pageview  (#772)
* DP-8453: Use TwigJS to lazy render results-headings to avoid HTTP request on every page

* Also convert js-pagination to lazy twig render

* Add changelog
