+++ styleguide/source/_patterns/02-molecules/results-heading.twig
@@ -1,6 +1,6 @@
 <div class="ma__results-heading js-results-heading">
   <div class="ma__results-heading__container">
-    {% set headingLevel = resultsHeading.level ? : 2 %}
+    {% set headingLevel = resultsHeading.level ? resultsHeading.level : 2 %}
     <h{{ headingLevel }} class="ma__results-heading__title">Showing {{ resultsHeading.numResults }} {% if resultsHeading.totalResults %} of {{ resultsHeading.totalResults }}{% endif %} results {% if resultsHeading.tags %}for:{% endif %}{% if resultsHeading.subject %}<span class="visually-hidden">for {{ resultsHeading.subject }}</span>{% endif %}</h{{ headingLevel }}>
     {% if resultsHeading.tags %}
       <div class="ma__results-heading__tags">
