+++ styleguide/source/assets/js/helpers/jQueryExtend_ajaxPattern.js
@@ -1,3 +1,5 @@
+import twiggy from './twiggy';
+
 export default function (window,document,$,undefined) {
   'use strict';
 
@@ -54,47 +56,12 @@ export default function (window,document,$,undefined) {
      *  The data structure which the pattern expects in order to render.
      */
     self.renderPattern = function(pattern,data) {
-      let template = Twig.twig({
-        href: self.getPatternPath(pattern),
-        id: pattern,
-        allowInlineIncludes: true,
-        // Set ma.patternPaths in your implementation's env.js file (see /source/_meta/_00-foot.twig)
-        namespaces: {
-          'base': ma.patternPaths['@base'],
-          'atoms': ma.patternPaths['@atoms'],
-          'molecules': ma.patternPaths['@molecules'],
-          'organisms': ma.patternPaths['@organisms'],
-          'templates': ma.patternPaths['@templates'],
-          'pages': ma.patternPaths['@pages'],
-          'meta': ma.patternPaths['@meta']
-        },
-        async: true, // Included/extended patterns loaded synchronously: https://github.com/twigjs/twig.js/issues/426
-        load: function(template) {
-          // Get compiled template with data.
-          let markup = template.render(data);
-          // Render markup in parent ajax-pattern component.
-          $element.html(markup);
-          // Trigger an event exposing the context for any listening pattern js module initialization.
-          $(document).trigger('ma:AjaxPattern:Render', [{'el': $element}]);
-        }
-      });
-    };
-
-    /**
-     * Returns the gettable path to a pattern based on its namespaced path.
-     *
-     * @param pattern
-     *   The namespaced path to a pattern (i.e. @atoms/decorative-link.twig)
-     *
-     * @returns {string}
-     *   The path to a pattern which can be used in a get request (i.e. /assets/patterns/atoms/decorative-link.twig )
-     */
-    self.getPatternPath = function(pattern) {
-      let pathParts = pattern.split("/");
-      let namespace = pathParts.shift();
-      let patternPath = pathParts.join("/");
-      // Set ma.patternPaths in your implementation's env.js file (see /source/_meta/_00-foot.twig)
-      return ma.patternPaths[namespace] + patternPath;
+      return twiggy(pattern)
+          .then(template => template.renderAsync(data))
+          .then(markup => {
+            $element.html(markup)
+            $(document).trigger('ma:AjaxPattern:Render', [{'el': $element}]);
+          })
     };
 
     /**
