--- styleguide/source/assets/js/templates/pagination.html
@@ -1,11 +1,0 @@
-<div class="ma__pagination__container">
-  <button class="ma__pagination__prev js-pagination-prev" type="button" {{# if prev.disabled }}disabled {{/if}}>{{ prev.text }}</button>
-  {{# each pages}}
-    {{# compare this.text "spacer" operator="==" }}
-      <span class="ma__pagination__spacer">&hellip;</span>
-    {{else}}
-      <button class="ma__pagination__page js-pagination-page {{# if this.active }}is-active{{/if}}" type="button" data-page="{{ this.text }}">{{ this.text }}</button>
-    {{/compare}}
-  {{/each}}
-  <button class="ma__pagination__next js-pagination-next" type="button" {{# if next.disabled }}disabled {{/if}}>{{ next.text }}</button>
-</div>
