--- styleguide/source/assets/js/templates/resultsHeading.html
@@ -1,21 +1,0 @@
-<div class="ma__results-heading__container">
-  <div class="ma__results-heading__title">Showing {{{numResults}}}
-    {{# if totalResults }}
-      of {{{totalResults}}}
-    {{/if}}
-    results
-    {{# if tags }}
-      for:
-    {{/if}}
-  </div>
-  {{# if tags }}
-  <div class="ma__results-heading__tags">
-    {{#each tags}}
-    <button type="button" class="ma__results-heading__tag js-results-heading-tag" data-ma-filter-type="{{{ this.type }}}" data-ma-filter-value="{{{ this.value }}}">{{{this.text}}}</button>
-    {{/each}}
-    {{# if tags }}
-    <button type="button" class="ma__results-heading__clear js-results-heading-clear">Clear all</button>
-    {{/if}}
-  </div>
-  {{/if}}
-</div>
